Updated dRally port.

+++ dRally/dRally.sh
@@ -28,12 +28,22 @@ $ESUDO chmod 666 log.txt
 export TERM=linux
 printf "\033c" > $CUR_TTY
 
-printf "\033c" > $CUR_TTY
 ## RUN SCRIPT HERE
+if [ -f "DeathRallyWin_10.exe" ]; then
+  echo "Extracting DeathRallyWin_10.exe" > $CUR_TTY
+  $EUSDO ./7z e -y DeathRallyWin_10.exe
+  $EUSDO rm -f DeathRallyWin_10.exe
+fi
+
+if [ ! -f "TR0.BPA" ]; then
+  echo "Game files missing, check README for installation instructions." > $CUR_TTY
+  sleep 5
+  exit
+fi
 
 echo "Starting game." > $CUR_TTY
 
-$GPTOKEYB "drally_linux" -c drally.gptk &
+$GPTOKEYB "drally_linux" -c drally.gptk textinput &
 $TASKSET ./drally_linux 2>&1 | $ESUDO tee -a ./log.txt
 
 $ESUDO kill -9 $(pidof gptokeyb)
